Use 64 values for epoch values.
git-svn-id: https://darwin.bournemouth.ac.uk/pdvrieze/svn/darwin/trunk/sql@291 3bae1aa3-1c52-4c45-a669-1fb9b31e7360

+++ darwin-sql/webauth/tokens.sql
@@ -6,7 +6,7 @@ CREATE TABLE `tokens` (
   `ip` varchar(24) NOT NULL,
   `keyid` int(11),
   `token` varchar(45) NOT NULL,
-  `epoch` int(11) NOT NULL,
+  `epoch` bigint NOT NULL,
   PRIMARY KEY (`tokenid`),
   FOREIGN KEY (`user`) REFERENCES `users` (`user`),
   FOREIGN KEY (`keyid`) REFERENCES `pubkeys` (`keyid`)
